requirements.txt
# Core Dependencies
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
python-dotenv==1.0.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# LINE Bot
line-bot-sdk==3.7.0

# Database & Caching
sqlalchemy==2.0.23
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Scheduling
schedule==1.2.0

# HTTP Client for APIs
httpx==0.25.2

# JSON Processing
ujson==5.8.0

# Logging
loguru==0.7.2

# Mistral AI Integration
mistralai>=1.9.0

# Web Search (Free - DuckDuckGo)
duckduckgo-search==8.1.1

# LINE Bot Integration
line-bot-sdk==3.7.0
requests>=2.31.0

